+++ packages/money-account-balance-service/src/response.types.ts
@@ -15,6 +15,16 @@ export type MusdEquivalentValueResponse = {
 };
 
 /**
+ * Response from {@link MoneyAccountBalanceService.getMoneyAccountBalance}.
+ */
+export type MoneyAccountBalanceResponse = {
+  musdBalance: string;
+  musdSHFvdValueInMusd: string;
+  // TODO: Add to function
+  totalBalance: string;
+};
+
+/**
  * Response from {@link MoneyAccountBalanceService.getVaultApy}.
  * All APY / fee values are decimals (multiply by 100 for percentage).
  *